more robust release script

+++ test/mk-release.sh
@@ -5,22 +5,32 @@
 # $1 is the tag to package
 
 if [ -z "${1}" ]; then echo "tag not given, exiting"; exit; fi
-echo tag to create release from: "$1"
+echo tag to create release from: \""$1"\"
+
+TAG=$(grep -P -o "^v\d+\.\d+\.\d+$" <<< "$1")
+if [[ "$TAG" == "" ]]; then
+    echo "TAG \"$1\" does not match required tag patten, exiting"
+    exit 1
+fi
 
 DIR=/tmp/dar-backup
 TARFILE="dar-backup-linux-${1}.tar.gz"
 
-rm -fr "$DIR"
-cd /tmp
+if [[ -e "$DIR" ]]; then rm -fr "$DIR" || exit 1; fi
+if [[ -f "/tmp/$TARFILE" ]]; then rm "/tmp/$TARFILE" || exit 1; fi
+
+cd /tmp || exit 1
 git clone https://github.com/per2jensen/dar-backup.git
-cd dar-backup
+cd dar-backup || exit 1
+
 git  checkout "tags/$1" -b "release-$1"
 chmod +x bin/install.sh
 rm -fr "$DIR/.git"
 rm -fr "$DIR/.github"
 rm -fr "$DIR/test"
+
 echo "This package is built from tag: $1" > VERSION
-cd $DIR/..
+cd $DIR/.. || exit 1
 tar czvf "$TARFILE" dar-backup
 echo SHA256:
 sha256sum "$TARFILE"
